+++ rio_tiler/reader.py
@@ -327,7 +327,7 @@ def point(
     post_process: Optional[
         Callable[[numpy.ndarray, numpy.ndarray], DataMaskType]
     ] = None,
-) -> List:
+) -> Tuple[List, List[str]]:
     """Read a pixel value for a point.
 
     Args:
@@ -344,6 +344,7 @@ def point(
 
     Returns:
         list: Pixel value per band indexes.
+        list: band names
 
     """
     if isinstance(indexes, int):
@@ -359,6 +360,8 @@ def point(
         raise PointOutsideBounds("Point is outside dataset bounds")
 
     indexes = indexes if indexes is not None else src_dst.indexes
+
+    band_names = [f"b{idx}" for idx in indexes]
 
     vrt_params = dict(add_alpha=True, resampling=Resampling[resampling_method])
     nodata = nodata if nodata is not None else src_dst.nodata
@@ -386,4 +389,4 @@ def point(
     if post_process:
         point_values, _ = post_process(point_values, mask)
 
-    return point_values.tolist()
+    return point_values.tolist(), band_names