+++ CHANGES.md
@@ -47,6 +47,63 @@
         >>> ["green_b1", "red_b1"]
     ```
 
+* depreciate `asset_expression` in MultiBaseReader. Use of expression is now possible
+* `expression` for MultiBaseReader must be in form of `{asset}_b{index}`
+
+    ```python
+    # before
+    with STACReader("stac.json") as stac:
+        stac.tile(701, 102, 8, expression="green/red")
+
+    # now
+    with STACReader("stac.json") as stac:
+        stac.tile(701, 102, 8, expression="green_b1/red_b1")
+    ```
+
+* `rio_tiler.reader.point()` (and all Reader's point methods) now return a **Tuple** of values and band names
+
+    ```python
+    # before
+    with rasterio.open("cog.tif") as src::
+        v = rio_tiler.reader.point(10.20, -42.0)
+        print(v)
+        >>> [0, 0, 0]
+
+    with COGReader("cog.tif") as cog:
+        print(cog.point(10.20, -42.0))
+        >>> [0, 0, 0]
+
+    # now
+    with rasterio.open("cog.tif") as src::
+        v, band_names = rio_tiler.reader.point(10.20, -42)
+        print(v)
+        >>> [0, 0, 0]
+        print(band_names)
+        >>> ["b1", "b2", "b3"]
+
+    with COGReader("cog.tif") as cog:
+        print(cog.point(10.20, -42.0))
+        >>> ([0, 0, 0], ["b1", "b2", "b3"])
+    ```
+
+* `MultiBaseReader.point()` method now returns data as flat (merged) list (instead of a list of list)
+
+    ```python
+    # before
+    with STACReader("stac.json") as stac:
+        pt = stac.point(10.20, -42, assets=("green", "red"))
+        print(pt)
+        >>> [[0], [0]]
+
+    # now
+    with STACReader("stac.json") as stac:
+        pt, names = stac.point(10.20, -42, assets=("green", "red"))
+        print(pt)
+        >>> [0, 0]
+        print(names)
+        >>> ["green_b1", "red_b1"]
+    ```
+
 # 3.1.6 (2022-07-22)
 
 * Hide `NotGeoreferencedWarning` warnings in `utils.render` and `utils.resize_array`